feat: adjust download speed
