+++ android/app/build.gradle
@@ -119,8 +119,8 @@ android {
         applicationId 'com.sigaa'
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
-        versionCode 40
-        versionName "1.1.0"
+        versionCode 41
+        versionName "1.1.1"
         python {
             buildPython "C:\\Python311\\python.exe"
         }
