+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "sigaa",
-  "version": "1.1.0",
+  "version": "1.1.1",
   "scripts": {
     "start": "expo start --dev-client",
     "android": "expo run:android",
@@ -23,7 +23,6 @@
     "expo-constants": "~14.2.1",
     "expo-file-system": "~15.2.2",
     "expo-intent-launcher": "~10.5.2",
-    "expo-linear-gradient": "~12.1.2",
     "expo-sharing": "~11.2.2",
     "expo-splash-screen": "~0.18.2",
     "expo-status-bar": "~1.4.4",
@@ -34,10 +33,8 @@
     "react-dom": "18.2.0",
     "react-native": "0.71.8",
     "react-native-easy-grid": "^0.2.2",
-    "react-native-easy-toast": "^2.3.0",
     "react-native-gesture-handler": "~2.9.0",
     "react-native-pager-view": "6.1.2",
-    "react-native-read-more-text": "^1.1.2",
     "react-native-render-html": "^6.3.4",
     "react-native-safe-area-context": "4.5.0",
     "react-native-screens": "~3.20.0",
