+++ src/api/donwloadDisciplina.ts
@@ -3,8 +3,9 @@ import * as cheerio from "cheerio";
 import { parse } from "node-html-parser";
 import { Alert, ToastAndroid } from "react-native";
 import getPermissions from "../hooks/getPermissions";
-import { replaceHeader, saveFile } from "./../utils/globalUtil";
+import { replaceHeader } from "./../utils/globalUtil";
 import { PythonModule } from "./api";
+import { saveFile } from "./files";
 
 export const donwloadDisciplina = async (json: any, javax: string) => {
   try {
@@ -18,7 +19,7 @@ export const donwloadDisciplina = async (json: any, javax: string) => {
     ToastAndroid.showWithGravity(
       "Baixando o arquivo, agurade um momento...",
       ToastAndroid.SHORT,
-      ToastAndroid.CENTER
+      ToastAndroid.BOTTOM
     );
 
     const response = String(
@@ -66,6 +67,11 @@ export const donwloadDisciplina = async (json: any, javax: string) => {
           );
         }
       }
+    } else {
+      Alert.alert(
+        "Erro",
+        "Erro ao baixar o arquivo, tente novamente mais tarde."
+      );
     }
   } catch (e) {
     Alert.alert(
