+++ src/api/downloadMenu.ts
@@ -3,66 +3,72 @@ import * as cheerio from "cheerio";
 import { parse } from "node-html-parser";
 import { Alert, ToastAndroid } from "react-native";
 import getPermissions from "../hooks/getPermissions";
-import { replaceHeader, saveFile } from "../utils/globalUtil";
+import { replaceHeader } from "../utils/globalUtil";
 import { PythonModule } from "./api";
+import { saveFile } from "./files";
 
 export const downloadMenu = async (payload: any, controller: any) => {
-  // try {
-  getPermissions();
-  ToastAndroid.showWithGravity(
-    "Baixando o arquivo, agurade um momento...",
-    ToastAndroid.SHORT,
-    ToastAndroid.CENTER
-  );
-  const response = String(
-    await PythonModule.download(
-      "https://sig.ifsudestemg.edu.br/sigaa/portais/discente/discente.jsf",
-      JSON.stringify(payload)
-    )
-  );
-  const regex = /'content': \[(.*?)\], 'headers': \{(.*?)\}/;
-  const matches: any = response.match(regex);
-  if (matches) {
-    const headersFormated = `{${matches[2].replace(/'/gm, '"')}}`.replace(
-      /"att.*""/gm,
-      replaceHeader
+  try {
+    getPermissions();
+    ToastAndroid.showWithGravity(
+      "Baixando o arquivo, agurade um momento...",
+      ToastAndroid.SHORT,
+      ToastAndroid.BOTTOM
     );
-    const headers = JSON.parse(headersFormated);
-    const content = JSON.parse(`{"contents": [${matches[1]}]}`);
-    if (
-      headers["Content-Disposition"] ||
-      headers["Content-disposition"] ||
-      headers["content-disposition"]
-    ) {
-      const file =
+    const response = String(
+      await PythonModule.download(
+        "https://sig.ifsudestemg.edu.br/sigaa/portais/discente/discente.jsf",
+        JSON.stringify(payload)
+      )
+    );
+    const regex = /'content': \[(.*?)\], 'headers': \{(.*?)\}/;
+    const matches: any = response.match(regex);
+    if (matches) {
+      const headersFormated = `{${matches[2].replace(/'/gm, '"')}}`.replace(
+        /"att.*""/gm,
+        replaceHeader
+      );
+      const headers = JSON.parse(headersFormated);
+      const content = JSON.parse(`{"contents": [${matches[1]}]}`);
+      if (
         headers["Content-Disposition"] ||
         headers["Content-disposition"] ||
-        headers["content-disposition"];
-      let type = headers["Content-Type"];
-      if (type === undefined) type = "application/octet-stream";
-      await saveFile(file, type, content.contents);
-    } else {
-      const $ = cheerio.load(
-        Buffer.from(content.contents, "binary").toString()
-      );
-      const turmas = parse($.html());
-      if (turmas.querySelector("ul.erros")) {
-        Alert.alert(
-          "Erro",
-          turmas.querySelector("ul.erros")?.textContent.trim()
-        );
+        headers["content-disposition"]
+      ) {
+        const file =
+          headers["Content-Disposition"] ||
+          headers["Content-disposition"] ||
+          headers["content-disposition"];
+        let type = headers["Content-Type"];
+        if (type === undefined) type = "application/octet-stream";
+        await saveFile(file, type, content.contents);
       } else {
-        Alert.alert(
-          "Erro",
-          "Erro ao baixar o arquivo, tente novamente mais tarde."
+        const $ = cheerio.load(
+          Buffer.from(content.contents, "binary").toString()
         );
+        const turmas = parse($.html());
+        if (turmas.querySelector("ul.erros")) {
+          Alert.alert(
+            "Erro",
+            turmas.querySelector("ul.erros")?.textContent.trim()
+          );
+        } else {
+          Alert.alert(
+            "Erro",
+            "Erro ao baixar o arquivo, tente novamente mais tarde."
+          );
+        }
       }
+    } else {
+      Alert.alert(
+        "Erro",
+        "Erro ao baixar o arquivo, tente novamente mais tarde."
+      );
     }
+  } catch (e) {
+    Alert.alert(
+      "Erro ao baixar o arquivo!",
+      "Provavelmente ele não está mais disponivel nos servidores do SIGAA!"
+    );
   }
-  // } catch (e) {
-  //   Alert.alert(
-  //     "Erro ao baixar o arquivo!",
-  //     "Provavelmente ele não está mais disponivel nos servidores do SIGAA!"
-  //   );
-  // }
 };
