+++ src/utils/globalUtil.ts
@@ -1,14 +1,7 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
-import { Buffer } from "buffer";
-import {
-  EncodingType,
-  StorageAccessFramework,
-  writeAsStringAsync,
-} from "expo-file-system";
-import { startActivityAsync } from "expo-intent-launcher";
-import { shareAsync } from "expo-sharing";
+
 import parse from "node-html-parser";
-import { Alert, Platform, ToastAndroid } from "react-native";
+import { ToastAndroid } from "react-native";
 
 export const formBody = (payload: any) =>
   Object.keys(payload)
@@ -29,73 +22,6 @@ export async function set() {
 export async function setTipoAluno(tipo = "graduacao") {
   await AsyncStorage.setItem("tipoAluno", tipo);
 }
-
-export const saveFile = async (file: string, type: string, data: string) => {
-  ToastAndroid.showWithGravity(
-    "Salvando o arquivo...",
-    ToastAndroid.SHORT,
-    ToastAndroid.CENTER
-  );
-  let local = await AsyncStorage.getItem("local");
-  if (local === null) {
-    const permissions =
-      await StorageAccessFramework.requestDirectoryPermissionsAsync();
-    if (!permissions.granted) {
-      return;
-    }
-    local = permissions.directoryUri;
-    await AsyncStorage.setItem("local", permissions.directoryUri);
-  }
-  StorageAccessFramework.createFileAsync(local, file, type)
-    .then(async (uri) => {
-      writeAsStringAsync(uri, Buffer.from(data, "binary").toString("base64"), {
-        encoding: EncodingType.Base64,
-      })
-        .then(() => {
-          let assetUriParts = uri.split("/");
-          let assetName = assetUriParts[assetUriParts.length - 1];
-          let url = `${local}/${assetName}`;
-          Alert.alert("Arquivo baixado com suceso!", "Deseja abrir ele?", [
-            {
-              text: "Cancelar",
-            },
-            {
-              text: "Abrir",
-              onPress: () => openFile(uri, type),
-            },
-          ]);
-        })
-        .catch(() => {
-          Alert.alert(
-            "Erro",
-            "Erro ao baixar o arquivo, tente novamente mais tarde."
-          );
-        });
-    })
-    .catch((e) => {});
-};
-
-const openFile = async (fileUri: string, type: string) => {
-  try {
-    if (Platform.OS === "android") {
-      await startActivityAsync("android.intent.action.VIEW", {
-        data: fileUri,
-        flags: 1,
-        type,
-      });
-    } else {
-      await shareAsync(fileUri, {
-        UTI: type,
-        mimeType: type,
-      });
-    }
-  } catch (error) {
-    Alert.alert(
-      "Erro",
-      "Não tem um aplicativo disponível para abrir esse arquivo."
-    );
-  }
-};
 
 export function isNumber(n: any) {
   return !isNaN(parseFloat(n)) && (isFinite(n) || Number.isInteger(n));
